requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.18.1
webdriver-manager==4.0.1
python-telegram-bot==20.7
python-dotenv==1.0.1
pytz==2024.1 